QUEUE-144: extract context listener write adapters
